ini2toml: get config types by inspection
The solution of recording these mappings in ini2toml was always fragile
and error-prone. I realized that I completely forgot to update it for
new services such as Linear and it turns out it's not too difficult to
get the types by introspection.

I'm fairly confident this doesn't break existing behavior because
`TestIni2Toml.test_bugwarriorrc` assures us that nothing changed in
running ini2toml on the example configuration.

This fixes the linear docs where I noticed the list-typed examples are
being converted to strings in the toml tabs.